+++ interceptor/intercept.py
@@ -3,64 +3,12 @@ import os
 import shutil
 import sys
 
-import pkg_resources
 from satella.coding import silence_excs
 from satella.files import write_to_file, read_in_file
 
 from interceptor.config import load_config_for, Configuration
+from interceptor.intercepting import intercept_tool, unintercept_tool, assert_intercepted, check
 from interceptor.whereis import filter_whereis
-
-INTERCEPTED = '-intercepted'
-INTERCEPTOR_WRAPPER_STRING = 'from interceptor.config import load_config_for'
-
-
-def intercept(tool_name: str) -> None:
-    source_file = pkg_resources.resource_filename(__name__, 'templates/cmdline.py')
-    try:
-        load_config_for(tool_name, None)
-    except KeyError:
-        print('No configuration found for %s, creating a default one' % (tool_name,))
-        shutil.copy(pkg_resources.resource_filename(__name__, 'templates/config'),
-                    os.path.join('/etc/interceptor.d', tool_name))
-    source = filter_whereis(sys.argv[1])
-    target_intercepted = source + INTERCEPTED
-    with silence_excs(UnicodeDecodeError), open(source, 'rb') as f_in:
-        if is_intercepted(tool_name):
-            print('Target is already intercepted. Aborting.')
-            sys.exit(1)
-    shutil.copy(source, target_intercepted)
-    os.unlink(source)
-    source_content = read_in_file(source_file, 'utf-8')
-    source_content = source_content.format(EXECUTABLE=sys.executable,
-                                           TOOLNAME=tool_name,
-                                           LOCATION=target_intercepted,
-                                           VERSION=pkg_resources.require('interceptor')[0].version)
-    write_to_file(source, source_content, 'utf-8')
-    os.chmod(source, 0o555)
-    print('Successfully intercepted %s' % (tool_name,))
-
-
-def is_intercepted(app_name: str) -> bool:
-    path = filter_whereis(app_name)
-
-    with silence_excs(UnicodeDecodeError), open(path, 'rb') as f_in:
-        intercepted_real = f_in.read(512).decode('utf-8')
-        return INTERCEPTOR_WRAPPER_STRING in intercepted_real
-
-
-def unintercept(app_name: str) -> None:
-    source = filter_whereis(app_name + INTERCEPTED)
-    src_name = source[:-len(INTERCEPTED)]
-    os.unlink(src_name)
-    shutil.move(source, src_name)
-    print('Successfully unintercepted %s' % (app_name,))
-    print('Leaving the configuration in place')
-
-
-def assert_intercepted(app_name: str) -> None:
-    if not is_intercepted(app_name):
-        print('%s is not intercepted' % (app_name,))
-        sys.exit(1)
 
 
 def banner():
@@ -69,6 +17,7 @@ def banner():
     * intercept undo foo - cancel intercepting foo
     * intercept configure foo - type in the configuration for foo in JSON format, end with Ctrl+D
     * intercept show foo - show the configuration for foo
+    * intercept status foo - display foo's status of interception and details about it's configuration
     * intercept display foo - enable displaying what is launched on foo's startup
     * intercept hide foo - disable displaying what is launched on foo's startup
     * intercept edit foo - launch a nano/vi to edit it's configuration
@@ -81,7 +30,10 @@ def banner():
     * intercept link foo bar - symlink bar's config file to that of foo
     * intercept copy foo bar - copy foo's configuration onto that of bar
     * intercept reset foo - reset foo's configuration (delete it and create a new one)
-    * intercept check foo - validate foo's configuration and reformat it
+    
+Use the optional switch --force is you need a command to complete despite the command telling you 
+that it is impossible to complete. One trick: already intercepted files won't be intercepted, because
+that would lead to overwriting of the original executable, so interceptor won't do that.
 ''')
 
 
@@ -91,7 +43,7 @@ def run():
         os.mkdir('/etc/interceptor.d')
 
     if len(sys.argv) == 2:
-        intercept(sys.argv[1])
+        intercept_tool(sys.argv[1])
     elif len(sys.argv) >= 3:
         op_name = sys.argv[1]
         app_name = sys.argv[2]
@@ -100,18 +52,7 @@ def run():
 
         if op_name == 'undo':
             assert_intercepted(app_name)
-            unintercept(app_name)
-        elif op_name == 'status':
-            if is_intercepted(app_name):
-                print('%s is intercepted' % (app_name,))
-                cfg_path = os.path.join('/etc/interceptor.d', app_name)
-                if os.path.islink(cfg_path):
-                    tgt_link = os.readlink(cfg_path).split('/')[-1]
-                    print('%s is scheduled to read configuration from %s' % (app_name, tgt_link))
-                else:
-                    print('%s has it\'s own configuration' % (app_name,))
-            else:
-                print('%s is NOT intercepted' % (app_name,))
+            unintercept_tool(app_name)
         elif op_name == 'configure':
             assert_intercepted(app_name)
             data = sys.stdin.read()
@@ -126,6 +67,8 @@ def run():
             assert_intercepted(app_name)
             config = read_in_file(os.path.join('/etc/interceptor.d', app_name), 'utf-8')
             print(config)
+        elif op_name == 'status':
+            check(app_name)
         elif op_name == 'edit':
             assert_intercepted(app_name)
             editor = filter_whereis('nano', abort_on_failure=False)
@@ -183,15 +126,8 @@ def run():
                 target = os.readlink(path).split('/')[-1]
                 print('%s config was a symlink to %s' % (app_name, target))
             os.unlink()
-            cfg = Configuration()
-            cfg.app_name = app_name
-            cfg.save()
+            Configuration(app_name=app_name).save()
             print('Configuration reset')
-        elif op_name == 'check':
-            assert_intercepted(app_name)
-            cfg = load_config_for(app_name, None)
-            cfg.save()
-            print('Configuration is valid')
         else:
             print('Unrecognized command %s' % (op_name,))
             banner()