+++ CHANGELOG.md
@@ -2,9 +2,11 @@ Changelog is normally kept [here](https://github.com/Cervi-Robotics/interceptor/
 
 This is only for new-version changes:
 
-# v1.3
+# v2.0
 
+* major refactor
 * `reset` will tell you if the configuration was a symlink
-
-
+* merged the role of `check` and `status` into a single `status` command
+* added the support for recognizing partial interceptions,
+  where not all executables of given file were patched